Updates to various subsystems
* Updated each subsystem so toplevel include file includes all required
subsystem files (for inclusion in other subsystems)
* Updated renderer subsystem to have lighting component
* Updated renderer subsystem with comments saying what should be
migrated to separate component
* Got rid of truncation and casting warnings

+++ Simulator/Simulator/Simulator.Shared/Engine/Engine.cpp
@@ -15,9 +15,9 @@ namespace Engine
 	{
 		mRenderer->Initialize();
 		mRenderer->SetCamera(
-			Vector<3, double>({ 5.0, 0.0, -5.0 }),						// position
-			Vector<3, double>({ 0.0, 0.0, 1.0 }),						// view
-			Vector<3, double>({ 0.0, 1.0, 0.0 }),						// up
+			Vector<3, float>({ 5.0, 0.0, -5.0 }),						// position
+			Vector<3, float>({ 0.0, 0.0, 1.0 }),						// view
+			Vector<3, float>({ 0.0, 1.0, 0.0 }),						// up
 			45.0f,														// FOVY
 			(float)(windowProperties.Width / windowProperties.Height),	// aspect ratio
 			1.0f,														// near plane
